put our Gsl::narrow in a separate file mimicing gsl/narrow
